+++ internal/integration/roundtrip.go
@@ -101,11 +101,16 @@ func (c *roundtripClient) MakeRequest(ctx context.Context, req *graphql.Request,
 	return nil
 }
 
-func newRoundtripClient(t *testing.T, endpoint string) graphql.Client {
+func newRoundtripClients(t *testing.T, endpoint string) []graphql.Client {
 	transport := &lastResponseTransport{wrapped: http.DefaultTransport}
-	return &roundtripClient{
-		wrapped:   graphql.NewClient(endpoint, &http.Client{Transport: transport}),
+	httpClient := &http.Client{Transport: transport}
+	return []graphql.Client{&roundtripClient{
+		wrapped:   graphql.NewClient(endpoint, httpClient),
 		transport: transport,
 		t:         t,
-	}
+	}, &roundtripClient{
+		wrapped:   graphql.NewClientUsingGet(endpoint, httpClient),
+		transport: transport,
+		t:         t,
+	}}
 }